Update functionnality of 'pst-node' package
The parameters 'angleA' and 'angleB' are the angle to place the line on the border of the node.
This is also the angle the line go out of the node.
Because this is 2 different informations, this modification of 'pst-node' create the 'outAngleA' and 'outAngleB' to independently change the second angle.