Add '\mytextsc' macro to fix '\myname'
The '\mytextsc' macro is fixing a problem with the Latin Modern fonts.
In fact, small caps sans serif are not defined in Latin Modern font which cause LaTeX use the serif font family to replace.
'\mytextsc' is using sans serif upper case in a smaller size to replace the small case.